Remove redundant isColorSupported function
It was only used on place, and it is much more reliable to test for the
existence of the variable 'T_USE_COLOR' than to execute the isColorSupported
function. This is because if the function would be called from a command
substitution then it would return false, even if it should return true.